Add a ssh script
I did this because I am really lazy and I always have to look this up when I am
setting up my computer. For now I have a script that creates the key if I don't
have one. If I do it just lists the items in the ~/.ssh directory. One thing
that was interesting to me is that I was not running the file using `source` so
the public key was not being added to the agent because it's in a different
process(or something like that). Now when it's ran by using `source` it
actually adds the key to the agent.Great stuff!