Add installation for ruby and zsh
This took a bit of work because I was struggling with getting ruby to install.
I went down a rabbit hole to figure out if it had to do with shell settings and
pathing issues. I don't think that was it. Turns out I was sending `set`
instead of the needed `sed` command to edit the ruby version I was getting back
from rbenv.

I also wanted to make this fancy. So I added ascii art to the top when it is
ran. So fancy!