Add vim configuration and set up to script
I did this because I want my vim set up and ready to go when I run my settop
script. This sources the file in my .dotfiles project that is installed by this
script. The vim_strap script should install and configure everything that I need
to get vim to work.