Update function names and create new functions
I did this because I wanted to output color that makes a little more sense and
dry up the script a bit by creating functions. I did this by creating a blue
colored print function for titles, and a yellow print as a type of 'warning'
print. This serves to print the things that are being skipped, while green will
be for actions that are actually taking place.

While doing this I noticed that some of the processes didn't have warning
scripts or used a bit of duplicated code (the brew installation for example). I
decided to pull as much of that into better functions for the script.