Add installation of a font in script
I did this because I want to be able to use this font in my terminal for all my
editing fun. This pulls the latest zip file from the repo and unzips it into my
users font directory. On a fresh laptop this directory will be empty and there
will be no files in it.