remove `command` from brew_install function
When running the `command -v brew ls...` part of the `brew_install` function it returns a value. It used to not. I wrote a small shell script to see that this should do the fix. There are comments on the web that talk about using the exit code but I couldn't get that to work properly. Since this works I am going to try this.

One thing that sucks is that I think this has something to do with a change to brew or maybe the new OS that I am running my script on. I wish I knew for sure which one it was. Brew didn't hit a major version but they did introduce a `--versions` (plural) so perhaps it's related to that. In any case, I'd just like to get this working in such a way that it is unlikely to break the next time I need it.