Add but comment out nvm for another time
I Did this because I want to install nvm through this bash script but it
looks like there is more to it to get it to work correctly. The command
-v nvm produces errors or, with modifications, produces the wrong result
to the conditional. I need more time so I'll update this later. For now
I'm leaving the commented code so I can spark my memory in the future.