Add xcode and command line tools installation
I did this because I want to install xcode on my machines for
development. This also includes the command line tools.

I explored some options for installing mac apps programmatically from
the command line and landed on the mas cli

https://github.com/mas-cli/mas

I use this for installing xcode from the mac app store.

I also added a note in the README because I need to make sure that the
computer is logged in correctly, or else mas will fail to install xcode.